feat: Add Run configuration for Cargo Command
A new Run configuration has been added for Cargo Command. This configuration allows running the project using the `cargo run` command. The working directory for this configuration is set to `file://$PROJECT_DIR$/Security`. The configuration also enables features, uses the STABLE channel, and requires sudo permission. Additionally, the output of the run command is redirected to the target directory of the Rust project in the Search Engine module.